Split the recording routes from their neighbours
src/web/recordings.rs had grown to the largest module under src/web/,
which was itself a split, and held three things that fail for different
reasons and answer to different callers.

The workers run on a timer against whatever the database holds and no
request drives them. The webhook answers LiveKit rather than a
candidate, so it authenticates on a signature instead of a session, and
everything below its handler bounds what a verified caller may then
touch. That is the reason it moves rather than its size: a trust
boundary is easier to review as a file than as the last third of one.

Start and stop stay with the reads. They looked like a fourth region by
line count and are not one: both take Owner, answer the account that
owns the recording, and share the same two helpers. What the move
leaves with one caller in its own file stops being visible crate-wide,
and the after sentinel, spelled at both replay routes, is spelled once.
